[chore](regression test) mv test_report_version_missing to p1 (#40184)
